+++ src/builder/index.ts
@@ -1,2 +1,4 @@
-export * from './typeorm-url.class';
+export * from './styles';
+export * from './styles.class';
+export * from './typeorm-uml.class';
 export * from './uml-builder.class';